--- packages/fast-components/src/tabs/tab-panel/tab-panel.ts
@@ -1,3 +1,0 @@
-import { FastElement } from "@microsoft/fast-element";
-
-export class TabPanel extends FastElement {}
