+++ packages/web-components/fast-components/src/tabs/fixtures/base.html
@@ -2,20 +2,20 @@
     <h1>Tabs</h1>
     <h4>Default</h4>
     <fast-tabs>
-        <fast-tab>Tab one</fast-tab>
-        <fast-tab>Tab two</fast-tab>
-        <fast-tab>Tab three</fast-tab>
-        <fast-tab-panel>
+        <fast-tab id="TabOne">Tab one</fast-tab>
+        <fast-tab id="TabTwo">Tab two</fast-tab>
+        <fast-tab id="TabThree">Tab three</fast-tab>
+        <fast-tab-panel id="TabPanelOne">
             Tab one content. This is for testing.
         </fast-tab-panel>
-        <fast-tab-panel>
+        <fast-tab-panel id="TabPanelTwo">
             Tab two content. This is for testing.
         </fast-tab-panel>
-        <fast-tab-panel>
+        <fast-tab-panel id="TabPanelThree">
             Tab three content. This is for testing.
         </fast-tab-panel>
     </fast-tabs>
-    <h4>Vertical</h4>
+    <!-- <h4>Vertical</h4>
     <fast-tabs orientation="vertical">
         <fast-tab>Tab one</fast-tab>
         <fast-tab>Tab two</fast-tab>
@@ -109,6 +109,6 @@
         <fast-tab-panel>
             Tab three content. This is for testing.
         </fast-tab-panel>
-    </fast-tabs>
+    </fast-tabs> -->
 
 </fast-design-system-provider>
