+++ packages/web-components/fast-components/src/tabs/tab-panel/tab-panel.styles.ts
