+++ packages/web-components/fast-components/src/tabs/tab-panel/tab-panel.template.ts
@@ -2,9 +2,7 @@ import { html } from "@microsoft/fast-element";
 import { TabPanel } from "./tab-panel";
 
 export const TabPanelTemplate = html<TabPanel>`
-    <template
-        role="tabpanel"
-    >
+    <template role="tabpanel">
         <slot></slot>
     </template>
 `;
