+++ packages/web-components/fast-components/src/tabs/tab-panel/tab-panel.ts
@@ -1,0 +1,3 @@
+import { FASTElement } from "@microsoft/fast-element";
+
+export class TabPanel extends FASTElement {}
