+++ packages/web-components/fast-components/src/tabs/tab/tab.styles.ts
