+++ packages/web-components/fast-components/src/tabs/tab/tab.template.ts
@@ -2,9 +2,7 @@ import { html } from "@microsoft/fast-element";
 import { Tab } from "./tab";
 
 export const TabTemplate = html<Tab>`
-    <template
-        slot="tab"
-    >
+    <template slot="tab">
         <slot></slot>
     </template>
 `;
