+++ packages/web-components/fast-components/src/tabs/tabs.styles.ts
